Visible Task Bar, proper slotting in window layouts of the windows system.
The MainWindow frame of the application now works much better but there are still 2 underlying issues -> the edges of the app are cut off a little when maximized and there are no grips to allow for custom resizing, just the presets. It's also impossible to drag the title bar to exit maximized app window, which could be a side effect of the edge being cut off, since it works in every other case, but could just be an oversight.